mv header to data
